requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

requests
dotenv

# Discord API
discord.py
